refactor[cm]: make generating result a static factory method
