+++ CodeModel/CodeParsers/CSharp/ClassCohesionAnalyzer/CohesivePartsOutput.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 
 namespace CodeModel.CodeParsers.CSharp.ClassCohesionAnalyzer
 {
@@ -15,7 +16,6 @@ namespace CodeModel.CodeParsers.CSharp.ClassCohesionAnalyzer
             TextsOfParts = textsOfTextsOfParts;
         }
 
-
         public override bool Equals(object obj)
         {
             if (obj is not CohesivePartsOutput cohesivePartsOutput) return false;
@@ -29,6 +29,69 @@ namespace CodeModel.CodeParsers.CSharp.ClassCohesionAnalyzer
         public override int GetHashCode()
         {
             return HashCode.Combine(AccessesToRemove, TextsOfParts);
+        }
+
+        public static CohesivePartsOutput[] GenerateOutput(CohesiveParts[] cohesiveParts, ResultMapper resultMapper)
+        {
+            if (cohesiveParts.Length == 0)
+                return Array.Empty<CohesivePartsOutput>();
+
+            var result = new CohesivePartsOutput[cohesiveParts.Length];
+            for (var i = 0; i < cohesiveParts.Length; i++)
+            {
+                var part = cohesiveParts[i];
+                var accessesToRemove = GetAccessesToRemoveText(part, resultMapper);
+                var textsOfParts = part.Parts.Select(p => GetClassPartText(p, resultMapper)).ToList();
+
+                result[i] = new CohesivePartsOutput(accessesToRemove, textsOfParts);
+            }
+
+            return result;
+        }
+
+        private static string GetAccessesToRemoveText(CohesiveParts part, ResultMapper resultMapper)
+        {
+            if (part.AccessesToRemove.Count == 0)
+                return "Class is already disconnected. No accesses should be removed.\n";
+
+            var builder = new StringBuilder();
+            builder.Append("To perform refactoring remove following method-field accesses:\n");
+            foreach (var access in part.AccessesToRemove)
+            {
+                var method = resultMapper.MethodsMapping[access.Method].Name;
+                var dataMember = resultMapper.FieldsMapping.ContainsKey(access.Field)
+                    ? resultMapper.FieldsMapping[access.Field].Name
+                    : resultMapper.AccessorsMapping[access.Field].Name;
+
+                builder.Append("Method: ");
+                builder.Append(method);
+                builder.Append(" -> Field: ");
+                builder.Append(dataMember);
+                builder.Append('\n');
+            }
+
+            return builder.ToString();
+        }
+
+        private static string GetClassPartText(ClassPart classPart, ResultMapper resultMapper)
+        {
+            var dataMembers = classPart.Accesses.GroupBy(access => access.Field).Select(group => group.Key).ToList();
+            var normalMethods = classPart.Accesses.GroupBy(access => access.Method).Select(group => group.Key).ToList();
+
+            var builder = new StringBuilder();
+            builder.Append("Cohesive part:\nFields & Accessors: ");
+            var fields = dataMembers.Where(dataMember => resultMapper.FieldsMapping.ContainsKey(dataMember))
+                .Select(i => resultMapper.FieldsMapping[i].Name);
+            var accessors = dataMembers.Where(dataMember => resultMapper.AccessorsMapping.ContainsKey(dataMember))
+                .Select(i => resultMapper.AccessorsMapping[i].Name);
+            builder.AppendJoin(", ", fields);
+            builder.AppendJoin(", ", accessors);
+
+            builder.Append("\nNormal methods: ");
+            var methods = normalMethods.Select(i => resultMapper.MethodsMapping[i].Name);
+            builder.AppendJoin(", ", methods);
+
+            return builder.ToString();
         }
     }
 }
